+++ src/vmm/src/vstate/vcpu/aarch64.rs
@@ -5,6 +5,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the THIRD-PARTY file.
 
+use kvm_bindings::*;
 use kvm_ioctls::*;
 use logger::{error, IncMetric, METRICS};
 use utils::vm_memory::{Address, GuestAddress, GuestMemoryMmap};
@@ -12,7 +13,8 @@ use versionize::{VersionMap, Versionize, VersionizeError, VersionizeResult};
 use versionize_derive::Versionize;
 
 use crate::arch::aarch64::regs::{
-    Aarch64RegisterOld, Aarch64RegisterRef, Aarch64RegisterVec, KVM_REG_ARM_TIMER_CNT,
+    arm64_core_reg_id, offset__of, Aarch64RegisterOld, Aarch64RegisterRef, Aarch64RegisterVec,
+    KVM_REG_ARM_TIMER_CNT,
 };
 use crate::arch::aarch64::vcpu::{
     get_all_registers, get_all_registers_ids, get_mpidr, get_mpstate, get_registers, set_mpstate,
@@ -208,11 +210,20 @@ impl KvmVcpu {
     pub fn dump_cpu_config(&self) -> Result<CpuConfiguration, KvmVcpuError> {
         let mut reg_list = get_all_registers_ids(&self.fd).map_err(KvmVcpuError::DumpCpuConfig)?;
 
+        let kvm_reg_pc = {
+            let kreg_off = offset__of!(kvm_regs, regs);
+            let pc_off = offset__of!(user_pt_regs, pc) + kreg_off;
+            arm64_core_reg_id!(KVM_REG_SIZE_U64, pc_off)
+        };
+
         // KVM_REG_ARM_TIMER_CNT should be removed, because it depends on the elapsed time and
         // the dumped CPU config is used to create custom CPU templates to modify CPU features
         // exposed to guests or ot detect CPU configuration changes caused by firecracker/KVM/
         // BIOS.
-        reg_list.retain(|&reg_id| reg_id != KVM_REG_ARM_TIMER_CNT);
+        // The value of program counter (PC) is determined by the given kernel image. It should not
+        // be overwritten by a custom CPU template and does not need to be tracked in a fingerprint
+        // file.
+        reg_list.retain(|&reg_id| reg_id != KVM_REG_ARM_TIMER_CNT && reg_id != kvm_reg_pc);
 
         let mut regs = Aarch64RegisterVec::default();
         get_registers(&self.fd, &reg_list, &mut regs).map_err(KvmVcpuError::DumpCpuConfig)?;